v2 client lookup in utils
